Establish the app's base dir, ID, and secret
All quart apps working with sessions need a session encryption secret,
as well as assurance that the session cookie is for that app (and not
another app on the same host). Thus, we establish an app ID based on the
physical disk location of this quart app, and assign a session crypt
token, either from existing cached value on disk, or a new one. The app
ID is deterministically generated based on the app's base dir.