Deal with the new ExceptionGroup, and reloading changes.
Introduced in 3.11, a gather() might throw multiple exceptions from
the various subtasks. We need to catch this situation, and see if any
of those exceptions are MustReloadError.

The reload/restart was moved from hypercorn to quart. We need to track
that change since we aren't using quart's core code to run an app.